* Fixed session timeout mistype
* Fixed SQL attributes constraints

* Implemented possibility to create Object Types and custom attributes (yet to link to creation of custom ConnectedObjects) => SQL Table + Classes + Service & Controller created

* Added MVP page for Object Type creation

* Removed age from User data (should not be stored but calculated with birth date)

* Fixed Security Config not implemented role hierarchy

* Fixed Ids written as Long in repositories when stored as Integer

* Translation of some comments

+++ cydomotix/src/main/java/com/example/cydomotix/Repository/HouseRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HouseRepository extends CrudRepository<House, Long>{
+public interface HouseRepository extends CrudRepository<House, Integer>{
 
     //
 }
