+++ cydomotix/src/main/java/com/example/cydomotix/Repository/RoomRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomRepository extends CrudRepository<Room, Long>{
+public interface RoomRepository extends CrudRepository<Room, Integer>{
 
     //
 }
